receiver/requirements.txt
psycopg2-binary==2.9.10
schedule==1.2.2
geoip2==4.8.1
requests==2.32.3
fastapi==0.115.6
uvicorn==0.34.0
